[Ranges]テストで警告が出る箇所を修正

+++ libs/ranges/test/src/adaptors/unit_test_ranges_drop_view.cpp
@@ -183,7 +183,7 @@ struct SizedViewWithUnsizedSentinel : hamon::ranges::view_base
 
 	constexpr iterator begin() const { return iterator{m_first}; }
 	constexpr sentinel end() const { return sentinel{iterator{m_last}}; }
-	constexpr std::size_t size() const { return m_last - m_first; }
+	constexpr std::size_t size() const { return static_cast<std::size_t>(m_last - m_first); }
 };
 static_assert( hamon::ranges::random_access_range_t<SizedViewWithUnsizedSentinel<int>>::value, "");
 static_assert( hamon::ranges::sized_range_t<SizedViewWithUnsizedSentinel<int>>::value, "");
